Ensure only one digitized transcription is store on data import
This fixes an issue identified on #161 where duplicate transcriptions
were showing up on some documents. Approximately 30 documents have
both an editor and translator recorded in the spreadsheet, and in
those cases we needed to update the logic to ensure that the
transcription text was only stored with a single edition.